refactor: Extract startReplication method
